introduce statistics impl template
